backend/requirements.txt
fastapi
uvicorn[standard]
faster-whisper
python-multipart
python-dotenv
google-generativeai